requirements.txt
streamlit
plotly
pandas
requests
fpdf2
